src/main/java/com/learning/microservices/userservice/controllers/UserController.java
package com.learning.microservices.userservice.controllers;

import com.learning.microservices.userservice.beans.User;
import com.learning.microservices.userservice.services.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
public class UserController {
    @Autowired
    private UserService userService;

    @GetMapping("/users/{userId}")
    public ResponseEntity<User> getUser(@PathVariable Long userId) {
        return new ResponseEntity<>(userService.getUserDetails(userId), HttpStatus.OK);
    }

    @GetMapping("/users")
    public ResponseEntity<List<User>> getUsers() {
        List<User> users = userService.getAllUsers();
        return new ResponseEntity<>(users, HttpStatus.OK);
    }

    @PostMapping("/users")
    public ResponseEntity<User> addUser(@RequestBody User userFromRequest) {
        User userCreated = userService.addUser(userFromRequest);
        return new ResponseEntity<>(userCreated, HttpStatus.CREATED);
    }
}
